添加文件权限管理及优化用户查询逻辑

+++ intelligent-office-ds/ds-api/src/main/java/com/thtf/chat/service/impl/BusUserInfoServiceImpl.java
@@ -185,7 +185,8 @@ public class BusUserInfoServiceImpl extends ServiceImpl<BusUserInfoMapper, BusUs
     @Override
     public RestResponse updateByUserId(UpdateUserInfoDto user) {
         try {
-            if (user.getRoleIds().size()>0) {
+//            分配角色
+            if (user.getRoleIds()!=null&&user.getRoleIds().size()>0) {
                 LambdaQueryWrapper<SysUserRoleEntity> queryWrapper = new LambdaQueryWrapper<>();
                 queryWrapper.eq(SysUserRoleEntity::getUserId, user.getId());
                 sysUserRoleRepo.remove(queryWrapper);
